Remove 'parsing' package. Move all used functions to gensim.utils. Fix imports (in py and ipynb).
