+++ gensim/test/test_parsing.py
@@ -8,8 +8,8 @@ Automated tests for the parsing module.
 import logging
 import unittest
 import numpy as np
-from gensim.parsing.preprocessing import remove_stopwords, strip_punctuation2, strip_tags, strip_short, strip_numeric, strip_non_alphanum, strip_multiple_whitespaces, split_alphanum, stem_text
-
+from gensim.utils import remove_stopwords, strip_punctuation, strip_tags, strip_short, strip_numeric, \
+    strip_non_alphanum, strip_multiple_whitespaces, split_alphanum, stem_text
 
 # several documents
 doc1 = """C'est un trou de verdure où chante une rivière,
@@ -36,7 +36,7 @@ occurrence of a collection of exact words,
 for many searching purposes, a little fuzziness would help. """
 
 
-dataset = [strip_punctuation2(x.lower()) for x in [doc1, doc2, doc3, doc4]]
+dataset = [strip_punctuation(x.lower()) for x in [doc1, doc2, doc3, doc4]]
 # doc1 and doc2 have class 0, doc3 and doc4 avec class 1
 classes = np.array([[1, 0], [1, 0], [0, 1], [0, 1]])
